+++ proton-mail-uniffi/CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - 2024-00-00
 
+### Added
+
+  - Added `Id` type to represent local IDs.
+
+### Changed
+
+  - Changed all local IDs from `u64` to `Id`.
+
 ## [0.11.11] - 2024-08-20
 
 ### Added